+++ services/has-many-associator.js
@@ -9,8 +9,12 @@ function HasManyAssociator(model, association, opts, params, data) {
         var associatedIds = _.map(data.data, function (value) {
           return value.id;
         });
+
+        // NOTICE: Deactivate validation to prevent potential issues with custom model validations.
+        //         In this case, the full record attributes are missing which may raise an
+        //         unexpected validation error.
         return record['add' + _.upperFirst(params.associationName)](
-          associatedIds);
+          associatedIds, { validate: false });
       });
   };
 }